+++ example/lib/examples/spritebatch_example.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart' hide Animation;
 import 'package:tremble/tremble.dart';
+import 'package:tremble_example/tex_enum.dart';
 
 class SpritebatchExample extends ScreenController {
   late SpriteBatch batch;
@@ -11,6 +12,10 @@ class SpritebatchExample extends ScreenController {
   @override
   Future<void> preload(UpdateCallback progress, VoidCallback done) async {
     batch = await SpriteBatch.fromGdxPacker("assets/sprites.atlas");
+
+    // You can get the enum string like this or just write to tex_enum.dart directly :)
+    //await Clipboard.setData(ClipboardData(text: batch.getEnum()));
+
     done();
   }
 
@@ -18,19 +23,19 @@ class SpritebatchExample extends ScreenController {
   void setup(BuildContext context, double width, double height) {
     sprites.addAll([
       Sprite(
-        texture: batch.getTexture(Tex.MysteryBlock),
+        texture: batch.getTexture(Tex.mysteryBlock),
         position: Vec2(width * 0.5 - 200, height * 0.5),
         originX: 0.5,
         originY: 1,
       ),
       Sprite(
-        texture: batch.getTexture(Tex.GroundBlock),
+        texture: batch.getTexture(Tex.groundBlock),
         position: Vec2(width * 0.5, height * 0.5),
         originX: 1,
         originY: 1,
       ),
       Animation(
-        animations: [batch.getAnimation(Tex.Mario_Big_Run, speed: 8)],
+        animations: [batch.getAnimation(Tex.marioBigRun, speed: 8)],
         position: Vec2(width * 0.5 + 200, height * 0.5),
       )
     ]);
@@ -71,60 +76,4 @@ class SpritebatchExample extends ScreenController {
   void dispose() {
     batch.dispose();
   }
-}
-
-// ignore_for_file: constant_identifier_names
-////////// Enum for the animations and texture in the atlas.
-
-enum TexType { texture, animation }
-
-enum Tex implements AssetName {
-  Brick(TexType.texture),
-  Bush1(TexType.texture),
-  Bush2(TexType.texture),
-  Bush3(TexType.texture),
-  Castle(TexType.texture),
-  Cloud1(TexType.texture),
-  Cloud2(TexType.texture),
-  Cloud3(TexType.texture),
-  Coin(TexType.texture),
-  Coin_Underground(TexType.texture),
-  EmptyBlock(TexType.texture),
-  Flag(TexType.texture),
-  FlagPole(TexType.texture),
-  Goomba_Flat(TexType.texture),
-  Goomba_Walk(TexType.animation),
-  GroundBlock(TexType.texture),
-  HardBlock(TexType.texture),
-  Hill1(TexType.texture),
-  Hill2(TexType.texture),
-  Koopa_Shell(TexType.texture),
-  Koopa_Walk(TexType.animation),
-  MagicMushroom(TexType.texture),
-  Mario_Big_Idle(TexType.texture),
-  Mario_Big_Jump(TexType.texture),
-  Mario_Big_Run(TexType.animation),
-  Mario_Big_Slide(TexType.texture),
-  Mario_Small_Death(TexType.texture),
-  Mario_Small_Idle(TexType.texture),
-  Mario_Small_Jump(TexType.texture),
-  Mario_Small_Run(TexType.animation),
-  Mario_Small_Slide(TexType.texture),
-  MysteryBlock(TexType.texture),
-  OneUpMushroom(TexType.texture),
-  PipeBottom(TexType.texture),
-  PipeConnection(TexType.texture),
-  PipeTop(TexType.texture),
-  Starman(TexType.texture),
-  UndergroundBlock(TexType.texture),
-  UndergroundBrick(TexType.texture),
-  fire(TexType.animation),
-  fire_flower(TexType.animation),
-  star_anim(TexType.animation);
-
-  final TexType type;
-  const Tex(this.type);
-
-  @override
-  String get assetName => name;
 }