+++ lib/render/tex_area.dart
@@ -19,7 +19,3 @@ class TexArea {
   @override
   String toString() => 'TexArea(rect: $rect, pageIndex: $pageIndex)';
 }
-
-abstract interface class AssetName {
-  String get assetName;
-}